Create IR eyes - initial line following code - motors stop when black is detected and move forward otherwise
Motors are 'LOW' when black is detected - the infrared wavelength is absorbed by the, in this instance electrical tape causing the respective IR eye to go report a 'HIGH' signal. This then sends a signal to the Arduino telling it that the black line has been detected and to adjust course by turning off the motor connected to the infrared eye that detected the signal change. This code causes the motors to stop when the IR wavelength is not returned, later on, PWM will be used to adjust the motor speed. 

Black electrical tape basically absorbers the whole IR light spectrum. When the sensors detect something under them, we have them move forward. When the sensors do not detect anything the vehicle will adjust course in order to return to both sensors detect something.